Extracts AutoFilled ParameterSource into its own file.
